requirements.txt
absl-py==0.8.0
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.2.10
astor==0.8.0
attrs==20.2.0
beautifulsoup4==4.9.1
bokeh==2.2.1
bs4==0.0.1
certifi==2019.6.16
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
conda==4.3.16
configparser==5.0.0
cryptography==3.3.1
cssselect==1.1.0
dask==2.25.0
defusedxml==0.6.0
distributed==2.25.0
Django==3.1
django-allauth==0.44.0
django-crispy-forms==1.10.0
django-newsletter==0.9.1
django-tinymce4-lite==1.8.0
djangorestframework==3.12.2
elasticsearch==7.10.1
facebook-scraper==0.2.10
fake-useragent==0.1.11
fsspec==0.8.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
HeapDict==1.0.1
html2text==2020.1.16
idna==2.8
importlib-metadata==1.7.0
iniconfig==1.0.1
ISStreamer==1.0.1
Jinja2==2.11.2
jsmin==2.2.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
ldif3==3.1.1
locket==0.2.0
lxml==4.5.2
Markdown==3.1.1
MarkupSafe==1.1.1
mod-wsgi==4.7.1
more-itertools==8.5.0
msgpack==1.0.0
mysqlclient==2.0.1
netaddr==0.7.19
numexpr==2.7.0
numpy==1.17.1
oauthlib==3.1.0
packaging==20.4
pandas==0.25.1
paratext-text-selector==0.3.3
parse==1.17.0
partd==1.1.0
pbkdf2==1.3
Pillow==7.2.0
pluggy==0.13.1
protobuf==3.9.1
psutil==5.7.2
py==1.9.0
pyaes==1.6.1
pyasn1==0.4.8
pycosat==0.6.3
pycparser==2.20
pyee==7.0.3
pygame==1.9.6
PyJWT==2.0.0
pyparsing==2.4.7
pyppeteer==0.2.2
PyQt3D==5.13.0
PyQt5==5.13.0
PyQt5-sip==4.19.18
pyquery==1.4.1
pytest==6.0.1
python-card-me==0.9.3
python-dateutil==2.8.0
python3-openid==3.2.0
pytz==2019.2
PyYAML==5.3.1
requests==2.7.0
requests-html==0.10.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.1.2
scapy==2.4.3
schedule==0.6.0
shortcuts==0.11.0
six==1.12.0
sniff-probe-req==0.5.1
sorl-thumbnail==12.7.0
sortedcontainers==2.2.2
soupsieve==2.0.1
sqlparse==0.3.1
surlex==0.2.0
tblib==1.7.0
telebot==0.0.3
Telethon==1.17.5
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
toml==0.10.1
toolz==0.10.0
tornado==6.0.4
tqdm==4.48.2
typing-extensions==3.7.4.3
tzlocal==2.1
unicodecsv==0.14.1
urllib3==1.25.10
utils==1.0.1
w3lib==1.22.0
websockets==8.1
Werkzeug==0.15.5
wifi==0.3.8
wrapt==1.11.2
zict==2.0.0
zipp==3.1.0
